code authorization

+++ yii/controllers/ProjectController.php
@@ -11,6 +11,7 @@ use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\db\Query;
 use Yii;
+use yii\web\ForbiddenHttpException;
 
 /**
  * ProjectController implements the CRUD actions for Project model.
@@ -42,13 +43,17 @@ class ProjectController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new ProjectSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if (Yii::$app->user->can('indexProject')) {
+            $searchModel = new ProjectSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            throw new ForbiddenHttpException('You do not have permission to list project.');
+        }
     }
 
     /**
@@ -59,9 +64,13 @@ class ProjectController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if (Yii::$app->user->can('detailProject')) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new ForbiddenHttpException('You do not have permission to detail a project.');
+        }
     }
 
     public function actionListUsersByProject($id)
@@ -84,37 +93,41 @@ class ProjectController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Project();
+        if (Yii::$app->user->can('createProject')) {
+            $model = new Project();
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                // Lấy danh sách người dùng được chọn
-                // $selectedUserIds = $this->request->post('selectedUserIds', []);
-                $selectedStaffIds = $this->request->post('Project')['staffIds'];
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    // Lấy danh sách người dùng được chọn
+                    // $selectedUserIds = $this->request->post('selectedUserIds', []);
+                    $selectedStaffIds = $this->request->post('Project')['staffIds'];
 
-                if (!empty($selectedStaffIds)) {
-                    foreach ($selectedStaffIds as $staffId) {
-                        // Tạo một bản ghi StaffProject cho mỗi staff được chọn
-                        $staffProject = new ProjectStaff();
-                        $staffProject->userId = $staffId;
-                        $staffProject->projectId = $model->id;
-                        $staffProject->save();
+                    if (!empty($selectedStaffIds)) {
+                        foreach ($selectedStaffIds as $staffId) {
+                            // Tạo một bản ghi StaffProject cho mỗi staff được chọn
+                            $staffProject = new ProjectStaff();
+                            $staffProject->userId = $staffId;
+                            $staffProject->projectId = $model->id;
+                            $staffProject->save();
+                        }
                     }
+
+                    return $this->redirect(['view', 'id' => $model->id]);
                 }
-
-                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                $model->loadDefaultValues();
             }
+
+            // Lấy danh sách người dùng để hiển thị trong form
+            $users = User::find()->all();
+
+            return $this->render('create', [
+                'model' => $model,
+                'users' => $users,
+            ]);
         } else {
-            $model->loadDefaultValues();
+            throw new ForbiddenHttpException('You do not have permission to create a project.');
         }
-
-        // Lấy danh sách người dùng để hiển thị trong form
-        $users = User::find()->all();
-
-        return $this->render('create', [
-            'model' => $model,
-            'users' => $users,
-        ]);
     }
 
     /**
@@ -126,38 +139,42 @@ class ProjectController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (Yii::$app->user->can('updateProject')) {
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            // Lấy danh sách người dùng được chọn
-            $selectedStaffIds = $this->request->post('Project')['staffIds'];
-            if (!empty($selectedStaffIds)) {
-                // Xóa các bản ghi ProjectStaff cũ cho dự án này
-                ProjectStaff::deleteAll(['projectId' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                // Lấy danh sách người dùng được chọn
+                $selectedStaffIds = $this->request->post('Project')['staffIds'];
+                if (!empty($selectedStaffIds)) {
+                    // Xóa các bản ghi ProjectStaff cũ cho dự án này
+                    ProjectStaff::deleteAll(['projectId' => $model->id]);
 
-                // Tạo bản ghi cho mỗi người dùng trong bảng ProjectStaff
-                foreach ($selectedStaffIds as $staffId) {
-                    $projectStaff = new ProjectStaff();
-                    $projectStaff->projectId = $model->id;
-                    $projectStaff->userId = $staffId;
-                    $projectStaff->save();
+                    // Tạo bản ghi cho mỗi người dùng trong bảng ProjectStaff
+                    foreach ($selectedStaffIds as $staffId) {
+                        $projectStaff = new ProjectStaff();
+                        $projectStaff->projectId = $model->id;
+                        $projectStaff->userId = $staffId;
+                        $projectStaff->save();
+                    }
                 }
+            
+                return $this->redirect(['view', 'id' => $model->id]);
             }
-        
-            return $this->redirect(['view', 'id' => $model->id]);
+
+            // Lấy danh sách người dùng để hiển thị trong form
+            $users = User::find()->all();
+
+            // Lấy danh sách người dùng đã tham gia dự án
+            $selectedUserIds = ProjectStaff::find()->select('userId')->where(['projectId' => $model->id])->column();
+
+            return $this->render('update', [
+                'model' => $model,
+                'users' => $users,
+                'selectedUserIds' => $selectedUserIds,
+            ]);
+        } else {
+            throw new ForbiddenHttpException('You do not have permission to update a project.');
         }
-
-        // Lấy danh sách người dùng để hiển thị trong form
-        $users = User::find()->all();
-
-        // Lấy danh sách người dùng đã tham gia dự án
-        $selectedUserIds = ProjectStaff::find()->select('userId')->where(['projectId' => $model->id])->column();
-
-        return $this->render('update', [
-            'model' => $model,
-            'users' => $users,
-            'selectedUserIds' => $selectedUserIds,
-        ]);
     }
 
     /**
@@ -169,9 +186,13 @@ class ProjectController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('deleteProject')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException('You do not have permission to delete a project.');
+        }
     }
 
     /**
